+++ example/lib/models/models.dart
@@ -1,0 +1,1 @@
+export './topic.dart';