+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Add dio 2.0 example
+* [Android] Fix json array data convert will crash
+* [iOS] Add `use_frameworks!` example (only cocoapods 1.5.3)
+
 ## 0.0.2
 
 * Add SharedPreferencesPlugin